fix(plugin): linting issues.
