+++ mcpgateway/plugins/__init__.py
@@ -13,8 +13,8 @@ Exposes core MCP Gateway plugin components:
 """
 
 from mcpgateway.plugins.framework.manager import PluginManager
-from mcpgateway.plugins.framework.models import PluginViolation, PluginViolationError
-from mcpgateway.plugins.framework.plugin_types import GlobalContext, PromptPosthookPayload, PromptPrehookPayload
+from mcpgateway.plugins.framework.models import PluginViolation
+from mcpgateway.plugins.framework.plugin_types import GlobalContext, PluginViolationError, PromptPosthookPayload, PromptPrehookPayload
 
 __all__ = [
     "GlobalContext",
